tests(e2e): rename from sign-in to auth
